requirements.txt
# Core Dependencies
python-dotenv==1.0.0
aiofiles==23.2.1
asyncio==3.4.3

# CLI Interface
rich==13.7.0
click==8.1.7
prompt-toolkit==3.0.43

# AI/LLM Libraries
openai==1.6.1
anthropic==0.8.1
perplexity-python==1.0.0  # Note: May need custom implementation

# CrewAI Framework (optional)
crewai==0.1.0
langchain==0.1.0
langchain-community==0.0.10

# Data Processing
pandas==2.1.4
numpy==1.26.2
pydantic==2.5.3

# Web & API
fastapi==0.108.0
uvicorn==0.25.0
requests==2.31.0
aiohttp==3.9.1

# Utilities
pyyaml==6.0.1
jinja2==3.1.2
python-multipart==0.0.6

# Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==7.0.0
mypy==1.7.1

# Optional: For advanced features
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1